Item Code
Problem: Mail accompanied by items can be duplicated

For example: A mail accompanied by 2 apples was sent to 3 other players.

Each player opened the mail and received 2 apples each, resulting in 6 apples in total.

To fix this, I have to add restriction to the mail that can only be sent and opened by 1 end user.